inputsource: optimize mheg5 response [2/2]
PD#SWPL-45928

Problem:
System response not quickly enough
too many time cost in data transforming
and drawing

Solution:
1.ARGB data send as int type, save the
time cost in data type convert
2.Rewrit overlayview draw function

Verify:
t5d

+++ logicdtvkit/src/main/java/org/droidlogic/dtvkit/DtvkitGlueClient.java
@@ -52,7 +52,7 @@ public class DtvkitGlueClient {
     }
 
         //native callback
-    public void notifySubtitleCallback(int width, int height, int dstx, int dsty, int dstwidth, int dstheight, byte[] data)
+    public void notifySubtitleCallback(int width, int height, int dstx, int dsty, int dstwidth, int dstheight, int[] data)
     {
          Log.d(TAG, "notifySubtitleCallBack received!!! width = " + width + ", heigth = " + height);
          if (mTarget != null) {
@@ -202,7 +202,7 @@ public class DtvkitGlueClient {
     }
 
     public interface OverlayTarget {
-        void draw(int src_width, int src_height, int dst_x, int dst_y, int dst_width, int dst_height, byte[] data);
+        void draw(int src_width, int src_height, int dst_x, int dst_y, int dst_width, int dst_height, int[] data);
     }
 
     public interface SubtitleListener {
